fix: 버킷 오브젝트 수정, 삭제되지 않는 버그 수정

+++ src/main/java/capstone/facefriend/bucket/BucketService.java
@@ -9,6 +9,7 @@ import capstone.facefriend.member.exception.member.MemberException;
 import capstone.facefriend.member.exception.member.MemberExceptionType;
 import capstone.facefriend.member.multipartFile.ByteArrayMultipartFile;
 import capstone.facefriend.member.service.dto.faceInfo.FaceInfoResponse;
+import capstone.facefriend.resume.domain.Resume;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
@@ -18,12 +19,15 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
 
+@Transactional
 @Slf4j
 @RequiredArgsConstructor
 @Service
@@ -32,34 +36,30 @@ public class BucketService {
     @Value("${spring.cloud.aws.s3.bucket}")
     private String bucketName;
 
-    @Value("${spring.cloud.aws.s3.default-profile}")
-    private String defaultProfileS3Url;
-
+    @Value("${spring.cloud.aws.s3.default-faceInfo-s3url}")
+    private String defaultFaceInfoS3url;
     @Value("${spring.cloud.aws.s3.origin-postfix}")
     private String originPostfix;
-
     @Value("${spring.cloud.aws.s3.generated-postfix}")
     private String generatedPostfix;
-
     @Value("${spring.cloud.aws.s3.resume-postfix}")
-    private String resumeInfix;
+    private String resumePostfix;
 
     private final AmazonS3 amazonS3;
     private final MemberRepository memberRepository;
-    private final FaceInfoRepository faceInfoRepository;
 
     // FaceInfo : origin 업로드 & generated 업로드
-    public FaceInfoResponse uploadOriginAndGenerated(
+    public List<String> uploadOriginAndGenerated(
             MultipartFile origin,
-            ByteArrayMultipartFile generated,
-            Long memberId
+            ByteArrayMultipartFile generated
     ) throws IOException {
         /** upload origin to s3 */
         // set metadata
         ObjectMetadata originMetadata = new ObjectMetadata();
         originMetadata.setContentLength(origin.getInputStream().available());
         originMetadata.setContentType(origin.getContentType());
-        String originObjectName = memberId + originPostfix;
+
+        String originObjectName = UUID.randomUUID() + originPostfix;
         amazonS3.putObject(
                 new PutObjectRequest(
                         bucketName,
@@ -76,7 +76,7 @@ public class BucketService {
         generatedMetadata.setContentLength(generated.getInputStream().available());
         generatedMetadata.setContentType(generatedMetadata.getContentType());
 
-        String generatedObjectName = memberId + generatedPostfix;
+        String generatedObjectName = UUID.randomUUID() + generatedPostfix;
         amazonS3.putObject(
                 new PutObjectRequest(
                         bucketName,
@@ -85,53 +85,53 @@ public class BucketService {
                         generatedMetadata
                 ).withCannedAcl(CannedAccessControlList.PublicRead)
         );
-        String generatedS3Url = amazonS3.getUrl(bucketName, generatedObjectName).toString();
+        String generatedS3url = amazonS3.getUrl(bucketName, generatedObjectName).toString();
 
-        // FaceInfo 저장
-        FaceInfo faceInfo = FaceInfo.builder()
-                .originS3url(originS3url)
-                .generatedS3url(generatedS3Url)
-                .build();
-        faceInfoRepository.save(faceInfo); //
-
-        // Member 최신화 후 저장
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberException(MemberExceptionType.NOT_FOUND));
-        member.setFaceInfo(faceInfo);
-        memberRepository.save(member);
-
-        return new FaceInfoResponse(originS3url, generatedS3Url);
+        return List.of(originS3url, generatedS3url);
     }
 
     // FaceInfo : origin 수정 -> generated 수정
-    public FaceInfoResponse updateOriginAndGenerated(
+    public List<String> updateOriginAndGenerated(
             MultipartFile origin,
             ByteArrayMultipartFile generated,
             Long memberId
     ) throws IOException {
-        deleteOriginAndGenerated(memberId); // 기존에 저장되어있던 사진 삭제
-        return uploadOriginAndGenerated(origin, generated, memberId); // 새로 사진 저장
+        Member member = findMemberById(memberId);
+
+        String originS3url = member.getFaceInfo().getOriginS3url();
+        String generatedS3url = member.getFaceInfo().getGeneratedS3url();
+
+        if (originS3url.equals(defaultFaceInfoS3url) || generatedS3url.equals(defaultFaceInfoS3url)) {
+            return uploadOriginAndGenerated(origin, generated);
+        }
+
+        deleteOriginAndGenerated(memberId);
+        return uploadOriginAndGenerated(origin, generated);
     }
 
     // FaceInfo : origin 삭제 -> generated 삭제
-    public FaceInfoResponse deleteOriginAndGenerated(
+    public String deleteOriginAndGenerated(
             Long memberId
     ) {
-        String originObjectName = memberId + originPostfix;
+        Member member = findMemberById(memberId);
+
+        String originS3url = member.getFaceInfo().getOriginS3url();
+        String originObjectName = originS3url.substring(originS3url.lastIndexOf("/") + 1);
         amazonS3.deleteObject(new DeleteObjectRequest(bucketName, originObjectName));
 
-        String generatedObjectName = memberId + generatedPostfix;
+        String generatedS3url = member.getFaceInfo().getGeneratedS3url();
+        String generatedObjectName = generatedS3url.substring(generatedS3url.lastIndexOf("/") + 1);
         amazonS3.deleteObject(new DeleteObjectRequest(bucketName, generatedObjectName));
 
-        return new FaceInfoResponse(defaultProfileS3Url, defaultProfileS3Url);
+        return defaultFaceInfoS3url;
     }
+
+
 
     // Resume : images 업로드
     public List<String> uploadResumeImages(
-            List<MultipartFile> images,
-            Long memberId
+            List<MultipartFile> images
     ) throws IOException {
-        int idx = 0;
         List<String> resumeImageS3urls = new ArrayList<>();
 
         for (MultipartFile image : images) {
@@ -139,7 +139,8 @@ public class BucketService {
             metadata.setContentLength(image.getInputStream().available());
             metadata.setContentType(image.getContentType());
 
-            String imageObjectName = memberId + resumeInfix + idx;
+            String imageObjectName = UUID.randomUUID() + resumePostfix;
+
             amazonS3.putObject(
                     new PutObjectRequest(
                             bucketName,
@@ -148,33 +149,35 @@ public class BucketService {
                             metadata
                     ).withCannedAcl(CannedAccessControlList.PublicRead)
             );
-
             resumeImageS3urls.add(amazonS3.getUrl(bucketName, imageObjectName).toString());
-            idx++;
         }
-
         return resumeImageS3urls;
     }
 
     // Resume : images 삭제 -> images 업로드
     public List<String> updateResumeImages(
             List<MultipartFile> images,
-            Long memberId,
-            int size
+            Resume resume
     ) throws IOException {
-        deleteResumeImages(memberId,size);
-        return uploadResumeImages(images, memberId);
+        deleteResumeImages(resume);
+        return uploadResumeImages(images);
     }
 
     // Resume : images 삭제
     public void deleteResumeImages(
-            Long memberId,
-            int size
+            Resume resume
     ) {
-        for (int idx = 0; idx < size; idx++) {
-            String imageObjectName = memberId + resumeInfix + idx;
-            amazonS3.deleteObject(new DeleteObjectRequest(bucketName, imageObjectName));
+        List<String> resumeImageS3urls = resume.getResumeImageS3urls();
+
+        for (String resumeImageS3url : resumeImageS3urls) {
+            String resumeImageObjectName = resumeImageS3url.substring(resumeImageS3url.lastIndexOf("/") + 1);
+            amazonS3.deleteObject(new DeleteObjectRequest(bucketName, resumeImageObjectName));
         }
+    }
+
+    private Member findMemberById(Long memberId) {
+        return memberRepository.findById(memberId)
+                .orElseThrow(() -> new MemberException(MemberExceptionType.NOT_FOUND));
     }
 }
